Extract shared Roslyn worker protocol
Move request framing, response parsing, protocol reads, and worker template
rendering out of the process host. Pass the locked worker output reader
explicitly so the protocol layer remains independent of host lifecycle state.
